+++ sources/ScreenshotAnnotator/Helpers/ImageExtensions.cs
@@ -5,10 +5,15 @@ namespace ScreenshotAnnotator.Helpers;
 
 static class ImageExtensions
 {
+    public static void SavePng(this Bitmap image, Stream stream)
+    {
+        image.Save(stream, PngBitmapEncoderOptions.Default);
+    }
+
     public static Stream ToStream(this Bitmap image)
     {
         var stream = new MemoryStream();
-        image.Save(stream);
+        image.SavePng(stream);
         stream.Position = 0;
         return stream;
     }